refactor(web): colocate modules and remove dead scaffolding
