+++ packages/web/src/__tests__/helpers/web-test-seams.ts
@@ -9,7 +9,7 @@ import {
   registerUseStartGoogleAuthorizationForTests,
   resetUseStartGoogleAuthorizationForTests,
   type UseStartGoogleAuthorization,
-} from "@web/auth/google/authorization/useStartGoogleAuthorization.registry";
+} from "@web/auth/google/authorization/useStartGoogleAuthorization";
 import {
   resetGoogleAvailabilityForTests,
   setGoogleAvailabilityForTests,
