--- packages/web/src/__tests__/utils/web.test.util.ts
@@ -1,8 +1,0 @@
-export const arraysAreEqual = (a: unknown[], b: unknown[]) => {
-  return (
-    Array.isArray(a) &&
-    Array.isArray(b) &&
-    a.length === b.length &&
-    a.every((val, index) => val === b[index])
-  );
-};
