--- packages/web/src/common/constants/auth.constants.ts
@@ -1,16 +1,0 @@
-import { z } from "zod";
-
-export const AuthStateSchema = z.object({
-  hasAuthenticated: z.boolean().default(false),
-  lastKnownEmail: z.string().email().optional(),
-  shouldPromptSignUpAfterAnonymousCalendarChange: z.boolean().default(false),
-});
-
-export type AuthState = z.infer<typeof AuthStateSchema>;
-
-export const DEFAULT_AUTH_STATE: AuthState = {
-  hasAuthenticated: false,
-  shouldPromptSignUpAfterAnonymousCalendarChange: false,
-};
-
-export const UNAUTHENTICATED_USER = "UNAUTHENTICATED_USER";
