+++ packages/web/src/common/constants/env.constants.test.ts
@@ -2,9 +2,7 @@ import { describe, expect, it } from "bun:test";
 
 process.env.PORT = "3000";
 
-const { getApiBaseUrl, isGoogleAuthConfigured } = await import(
-  "./env.constants"
-);
+const { getApiBaseUrl } = await import("./env.constants");
 
 describe("getApiBaseUrl", () => {
   it("defaults to the local backend API using the configured port", () => {
@@ -21,19 +19,5 @@ describe("getApiBaseUrl", () => {
     expect(() => getApiBaseUrl()).toThrow(
       "PORT is required when API_BASEURL is not configured",
     );
-  });
-});
-
-describe("isGoogleAuthConfigured", () => {
-  it("rejects missing or empty Google client IDs", () => {
-    expect(isGoogleAuthConfigured()).toBe(false);
-    expect(isGoogleAuthConfigured("")).toBe(false);
-    expect(isGoogleAuthConfigured("undefined")).toBe(false);
-  });
-
-  it("accepts a custom Google client ID", () => {
-    expect(
-      isGoogleAuthConfigured("1234567890-example.apps.googleusercontent.com"),
-    ).toBe(true);
   });
 });
