+++ packages/web/src/components/CommandPalette/event.cmd.constants.ts
