+++ packages/web/src/components/CommandPalette/navigation.cmd.constants.ts
