--- packages/web/src/components/SpaceCharacter/SpaceCharacter.tsx
@@ -1,1 +1,0 @@
-export const SpaceCharacter = () => <>&nbsp;</>;
