--- packages/web/src/components/SpaceCharacter/index.ts
@@ -1,3 +1,0 @@
-import { SpaceCharacter } from "./SpaceCharacter";
-
-export { SpaceCharacter };
