+++ packages/web/src/events/types/local-event.record.ts
