+++ packages/web/src/views/GoogleAuthCallback/GoogleAuthCallback.tsx
@@ -6,9 +6,9 @@ import { useCompleteAuthentication } from "@web/auth/compass/hooks/useCompleteAu
 import { session } from "@web/auth/compass/session/Session";
 import { refreshUserMetadata } from "@web/auth/compass/user/util/user-metadata.util";
 import { completeGoogleAuthorization } from "@web/auth/google/authorization/complete-google-authorization";
-import { releaseNotesPromptActions } from "@web/auth/state/release-notes-prompt.store";
 import { showErrorToast } from "@web/common/utils/toast/error-toast.util";
 import { OverlayPanel } from "@web/components/OverlayPanel/OverlayPanel";
+import { releaseNotesPromptActions } from "@web/components/ReleaseNotesPrompt/release-notes-prompt.store";
 import { eventQueryKeys } from "@web/events/queries/event.query.keys";
 
 type CompleteAuthentication = ReturnType<typeof useCompleteAuthentication>;
